working version (a lot is still missing)

+++ include/core/Session.hpp
@@ -10,20 +10,20 @@ public:
 
   bool getActiveStatus() const;
   Subject getSubject() const;
-  int getDate() const;
+  long long getDate() const;
   int getGoalDuration() const;
   int getDuration() const;
 
   void setActiveStatus(bool active);
   void setSubject(Subject subject);
-  void setDate(int date);
+  void setDate(long long date);
   void setGoalDuration(int goal_duration);
   void setDuration(int duration);
 
 private:
   bool active_ = true;
   Subject subject_;
-  int date_ = 0;
+  long long date_ = 0;
   int goal_duration_ = 45;
   int duration_ = 0;
 };
